% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{Cov.Vars}
\alias{Cov.Vars}
\title{Function determining weights on covariates by their standard deviation}
\usage{
Cov.Vars(treated.covariates, donors.covariates)
}
\description{
Function determining weights on covariates by their standard deviation
}
