% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cross-validation_Code_byK.R
\name{cv_estimator}
\alias{cv_estimator}
\title{Solve for optimal model averaging parameter and associated weights and errors conditional on matching estimator.}
\usage{
cv_estimator(data, tune_pars, estimator)
}
\arguments{
\item{data}{A \code{list} containing three named elements: \describe{
\item{donors:}{A \eqn{TxN} matrix of outcome paths for untreated units, each column being a control unit.}
\item{treated:}{A \eqn{Tx1} matrix of outcome paths for the treated unit.}
\item{treatment:}{An integer. The period T' in which treatment occurs (\eqn{T'<T}).}
}}

\item{tune_pars}{A \code{list} containing 3 elements. You must specify the first, and you must specify
exctly one of the remaining two elements: \describe{
\item{m:}{ a vector of integers representing the set of candidate nearest neighbor estimators.}
\item{min_preperiods:}{an integer. The smallest number of periods allowed in a fold used for cross-validation.}
\item{set_f:}{a vector of integers. Identifies the set of folds used for cross-validation (see section 3 of paper).}
}}

\item{estimator}{A function defining an objective to minimize with respect to the tuning parameter values. Defaults
to the masc estimator (\link{solve_masc}), but one could in principle replace this with another estimator taking the same arguments.}
}
\value{
returns a list containing five objects:
\describe{
\item{phi:}{selected value for the model averaging parameter (1 is pure matching, 0 pure synthetic control).}
\item{m:}{selected matching (nearest neighbor) estimator.}
\item{weights:}{The vector length N containing weights placed on each control unit.}
\item{pred.error:}{The vector of treatment effects implied by the masc counterfactual, for periods T' to T.}
\item{cv.error:}{The cross-validation error value returned by the optimal set of tuning parameter values.}
}
}
\description{
Implements the matching and synthetic control (masc) estimator of Kellogg, Mogstad,
 Pouliot, and Torgovitsky (2019), \emph{conditional on a given matching estimator}.
 \link{masc} loops over evaluations of this function for each candidate matching estimator,
 and selects the one which minimizes cross-validation error.
}
\references{
Kellogg, M., M. Mogstad, G. Pouliot, and A. Torgovitsky. Combining Matching and Synthetic Control to Trade
 off Biases from Extrapolation and Interpolation. Working Paper, 2019.
}
\seealso{
Other masc functions: \code{\link{masc}},
  \code{\link{solve_masc}}
}
\concept{masc functions}
