% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotelgluth2024}
\alias{hotelgluth2024}
\title{Hotel Choice Experiment Dataset from Gluth et al. (2024)}
\format{
\subsection{\code{hotelgluth2024}}{

A data frame with multiple rows per subject, each representing a single trial
in a hotel choice experiment with three attributes:
\describe{
\item{dataset}{Identifier for the dataset type, set to "hotel"}
\item{subject}{Unique identifier for each participant in the experiment}
\item{trial}{Trial number within each subject's experimental session}
\item{opt1_att1}{Value of the first attribute for the first hotel option (distance to beach)}
\item{opt1_att2}{Value of the second attribute for the first hotel option (room size)}
\item{opt1_att3}{Value of the third attribute for the first hotel option (number of stars)}
\item{opt2_att1}{Value of the first attribute for the second hotel option (distance to beach)}
\item{opt2_att2}{Value of the second attribute for the second hotel option (room size)}
\item{opt2_att3}{Value of the third attribute for the second hotel option (number of stars)}
\item{difficulty}{Difficulty level of the choice trial, based on the difference in option ratings}
\item{choice}{Chosen hotel option (1 or 2)}
\item{rt}{Reaction time for the choice}
\item{att_w1}{Weight of the first attribute (distance to beach)}
\item{att_w2}{Weight of the second attribute (room size)}
\item{att_w3}{Weight of the third attribute (number of stars)}
\item{sigma}{Model parameter related to sampling noise}
\item{alpha}{Model parameter related to search rule sensitivity}
\item{delta}{Model parameter related to choice threshold}
}
}
}
\source{
Gluth, S., Deakin, J., & Rieskamp, J. (2024). A Theory of Multi-Attribute Search and Choice.
Unpublished manuscript, University of Hamburg and University of Basel.
}
\usage{
hotelgluth2024
}
\description{
A dataset from a multi-attribute decision-making experiment examining how people
search for and evaluate information when choosing between hotel options.
}
\details{
This dataset captures individual decision-making processes in a hotel choice experiment.
Participants were asked to choose between two hotel options described by three attributes:
distance to beach, room size, and number of stars. The dataset includes both the raw choice
data and estimated model parameters from the Multi-Attribute Search and Choice (MASC) model.
}
\references{
Gluth, S., Deakin, J., & Rieskamp, J. (2024). A Theory of Multi-Attribute Search and Choice.
Unpublished manuscript, University of Hamburg and University of Basel.
}
\keyword{datasets}
