% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masc.R
\name{MASC}
\alias{MASC}
\title{MASC - Mixed effect modeling of Associations of Single Cells}
\usage{
MASC(data, cluster, contrast, random.effects, fixed.effects, verbose = TRUE)
}
\arguments{
\item{data}{A data frame containing the contrast factor, random, and fixed effects for the model}

\item{cluster}{A factor indicating cluster assignments for each cell}

\item{contrast}{A vector indicating the factor to be used as a contrast term}

\item{random.effects}{A vector indicating which terms should be modeled as random effects covariates}

\item{fixed.effects}{A vector indicating wich terms should be modeled as fixed effects covariates}

\item{verbose}{TRUE/FALSE}
}
\value{
data frame containing calculated association p-values and odds ratios for each cluster tested
}
\description{
MASC - Mixed effect modeling of Associations of Single Cells
}
\examples{
# Create test dataset with three clusters of 100 cells each
test.df <- data.frame(cluster = factor(rep(c(1, 2, 3), each = 100)))
# Create 6 donors that are cases or controls and include covariates
donors.df <- data.frame(donor = rep(paste("Donor", LETTERS[1:6], sep = "_"), each = 50),
sex = rep(c("M", "F", "M", "F", "F", "M"), each = 50),
status = rep(c("Case", "Case", "Control", "Control", "Case", "Control"), each = 50))
# Now make cluster 1 mostly case, cluster 2 mostly controls, etc
cases <- donors.df[donors.df$status == "Case",]
cases <- cases[sample(nrow(cases)),]
controls <- donors.df[donors.df$status == "Control",]
controls <- controls[sample(nrow(controls)),]
test.df <- cbind(rbind(cases[1:75,], controls[1:25,], cases[76:115,], controls[26:85,], cases[116:150,], controls[86:150,]), test.df)
# Test set call
library(lme4)
MASC(data = test.df, cluster = test.df$cluster, contrast = "status", random.effects = "donor", fixed.effects = "sex")

}
