% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{NearestNeighbors}
\alias{NearestNeighbors}
\title{Nearest Neighbor selection on covariates or outcome paths
Defaults to unweighted distance on outcome paths if covariates are not specified,
or standard deviations with respect to covariates if covariates are specified.}
\usage{
NearestNeighbors(
  treated,
  donors,
  treated.covariates,
  donors.covariates,
  treatment,
  tune.pars,
  ...
)
}
\description{
Nearest Neighbor selection on covariates or outcome paths
Defaults to unweighted distance on outcome paths if covariates are not specified,
or standard deviations with respect to covariates if covariates are specified.
}
