% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{sc_estimator}
\alias{sc_estimator}
\title{solving for standard synthetic control weights}
\usage{
sc_estimator(donors, treated, treatment, nogurobi)
}
\arguments{
\item{donors}{A \eqn{TxN} matrix of outcome paths for untreated units, each column being a control unit.}

\item{nogurobi}{A logical value. If true, uses \link[LowRankQP]{LowRankQP} to solve the synthetic control estimator,
rather than \code{gurobi}.}

\item{treated:}{A \eqn{Tx1} matrix of outcomes for the treated unit.}

\item{treatment:}{An integer. The period T' in which forecasting begins (either the true treatment
period or the first period after a cross-validation fold).}
}
\value{
A list. The named component \code{weights.sc} contains the vector of synthetic control weights.
 Weights are ordered in the same manner as the columns in \code{donors}.
 The \code{objval.sc} component contains the objective value (pre-period fit) of the synthetic control.
}
\description{
Solves for synthetic control weights up to the designated treatment period.
}
\references{
Kellogg, M., M. Mogstad, G. Pouliot, and A. Torgovitsky. Combining Matching and Synthetic Control to Trade
 off Biases from Extrapolation and Interpolation. Working Paper, 2019.
}
\seealso{
Other masc functions: \code{\link{cv_masc}},
  \code{\link{masc_by_phi}}, \code{\link{masc}},
  \code{\link{solve_masc}}
}
\concept{masc functions}
