% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{solve_masc}
\alias{solve_masc}
\title{solving for matching and synthetic control weights}
\usage{
solve_masc(donors, treated, treatment, sc_est, tune_pars,
  nogurobi = FALSE)
}
\arguments{
\item{donors}{A \eqn{TxN} matrix of outcome paths for untreated units, each column being a control unit.}

\item{tune_pars}{A \code{list} containing one element:
\describe{\item{m:}{an integer identifying the candidate matching (nearest neighbor) estimator.}}}

\item{treated:}{A \eqn{Tx1} matrix of outcomes for the treated unit.}

\item{treatment:}{An integer. The period T' in which forecasting begins (either the true treatment
period or the first period after a cross-validation fold).

#'@param sc_est A \code{function} which constructs weights associated with a synthetic control-type estimator. See
\link{sc_estimator} for input and output if you'd prefer to substitute your own estimator.}
}
\value{
A list containing the weights associated with the two estimators. Weights are ordered
in the same manner as the columns in \code{donors}. The \code{weights.sc} and \code{weights.match}
named components contain the vector of synthetic control and nearest neighbor weights respectively.
 The \code{objval.sc} component contains the objective value (pre-period fit) of the synthetic control.
}
\description{
Solves separately for synthetic control weights and matching weights for a given
matching estimator, using data up to the treatment period.
}
\references{
Kellogg, M., M. Mogstad, G. Pouliot, and A. Torgovitsky. Combining Matching and Synthetic Control to Trade
 off Biases from Extrapolation and Interpolation. Working Paper, 2019.
}
\seealso{
Other masc functions: \code{\link{cv_masc}},
  \code{\link{masc_by_phi}}, \code{\link{masc}},
  \code{\link{sc_estimator}}
}
\concept{masc functions}
