% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masc.R
\name{generate_attributes}
\alias{generate_attributes}
\title{Generate Non-dominated Attribute Values}
\usage{
generate_attributes(n, m, lambda)
}
\arguments{
\item{n}{Integer. Number of options.}

\item{m}{Integer. Number of attributes per option.}

\item{lambda}{Numeric. Precision parameter for attribute value generation.}
}
\value{
Matrix of attribute values (n x m).
}
\description{
Internal function that generates attribute values for options ensuring no option
dominates another (i.e., is better on all attributes).
}
\keyword{internal}
