% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masc-package.R
\docType{package}
\name{masc-package}
\alias{masc}
\alias{masc-package}
\title{masc: Multi-Attribute Search and Choice Model}
\description{
Implementation of the Multi-Attribute Search and Choice (MASC) model for
decision making. This model simulates how people make decisions when comparing
options with multiple attributes by sequentially sampling information about
different attributes until reaching a decision.
}
\details{
The main function provided by this package is:
\itemize{
\item \code{\link{rMASC}}: Generate choices using the MASC model
}
}
\references{
Gluth, S., Deakin, J., & Rieskamp, J. (2026). A theory of multiattribute search
and choice. \emph{Psychological Review}. \doi{10.1037/rev0000614}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kiante-fernandez/masc}
  \item Report bugs at \url{https://github.com/kiante-fernandez/masc/issues}
}

}
\author{
\strong{Maintainer}: Kiante Fernandez \email{kiantefernan@gmail.com} (\href{https://orcid.org/0000-0002-8493-880X}{ORCID}) [copyright holder]

}
\keyword{internal}
