% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masc.R
\name{MASC}
\alias{MASC}
\title{MASC - Mixed effect modeling of Associations of Single Cells}
\usage{
MASC(dataset, cluster, contrast, random_effects = NULL,
  fixed_effects = NULL, verbose = FALSE, save_models = FALSE,
  save_model_dir = NULL)
}
\arguments{
\item{dataset}{A data frame containing the contrast factor, random, and fixed effects for the model}

\item{cluster}{A factor indicating cluster assignments for each cell}

\item{contrast}{A vector indicating the variable to be tested for association with cluster abundance. Must match a column in dataset.}

\item{random_effects}{A vector indicating which terms should be modeled as random effects covariates. Terms listed must match columns in dataset.}

\item{fixed_effects}{A vector indicating which terms should be modeled as fixed effects covariates. Terms listed must match columns in dataset.}

\item{verbose}{TRUE/FALSE}

\item{save_models}{Should MASC save the mixed-effects model objects generated for each cluster?}

\item{save_model_dir}{Location to save mixed-effect model objects. Defaults to current working directory.}
}
\value{
data frame containing calculated association p-values and odds ratios for each cluster tested
}
\description{
MASC - Mixed effect modeling of Associations of Single Cells
}
\examples{
# Create test dataset with three clusters of 100 cells each
test.df <- data.frame(cluster = factor(rep(c(1, 2, 3), each = 100)))
# Create 6 donors that are cases or controls and include covariates
donors.df <- data.frame(
  donor = rep(paste("Donor", LETTERS[1:6], sep = "_"), each = 50),
  sex = rep(c("M", "F", "M", "F", "F", "M"), each = 50),
  status = rep(c("Case", "Case", "Control", "Control", "Case", "Control"), each = 50)
)
# Now make cluster 1 mostly case, cluster 2 mostly controls, etc
cases <- donors.df[donors.df$status == "Case", ]
cases <- cases[sample(nrow(cases)), ]
controls <- donors.df[donors.df$status == "Control", ]
controls <- controls[sample(nrow(controls)), ]
test.df <- cbind(rbind(cases[1:75, ], controls[1:25, ], cases[76:115, ], controls[26:85, ], cases[116:150, ], controls[86:150, ]), test.df)
# Test set call
library(lme4)
MASC(dataset = test.df, cluster = test.df$cluster, contrast = "status", random_effects = "donor", fixed_effects = "sex")
}
