% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimator_Code.R
\name{solve_masc}
\alias{solve_masc}
\title{solving for matching and synthetic control weights}
\usage{
solve_masc(data, tune_pars, ...)
}
\arguments{
\item{data}{A \code{list} containing three named elements: \describe{
\item{donors:}{A \eqn{TxN} matrix of outcome paths for untreated units, each column being a control unit.}
\item{treated:}{A \eqn{Tx1} matrix of outcome paths for the treated unit.}
\item{treatment:}{An integer. The period T' in which treatment occurs (\eqn{T'<T}).}
}}

\item{tune_pars}{A \code{list} containing one element:
\describe{\item{m:}{an integer identifying the candidate matching (nearest neighbor) estimator.}}}
}
\value{
A list containing the weights associated with the two estimators. Weights are ordered
in the same manner as the columns in \code{data}. The \code{weights.sc} and \code{weights.match}
named components contain the vector of synthetic control and nearest neighbor weights respectively.
}
\description{
Solves separately for synthetic control weights and matching weights for a given
matching estimator, using information in \code{data} up to the treatment period
designated in that object.
}
\references{
Kellogg, M., M. Mogstad, G. Pouliot, and A. Torgovitsky. Combining Matching and Synthetic Control to Trade
 off Biases from Extrapolation and Interpolation. Working Paper, 2019.
}
\seealso{
Other masc functions: \code{\link{cv_estimator}},
  \code{\link{masc}}
}
\concept{masc functions}
